+++ test/unit/markable_test.rb
@@ -65,5 +65,23 @@ module Stamina
       assert_nil(@loaded["rr"])
       assert_equal([:loaded_pair,[:myownkey,nil,"myownkey"],:loaded_pair,[12,nil,:twelve],:loaded_pair,[:myownkey,"myownkey",36]],@loaded.changes)
     end
+
+    def test_marks
+      markable = MyMarkableObject.new
+      markable[:hello] = "world"
+      markable[:foo] = "bar"
+      assert_equal [], markable.marks
+      assert_equal ["world"], markable.marks(:hello)
+      assert_equal ["bar", "world"], markable.marks(:foo, :hello)
+      assert_equal [nil, "world"], markable.marks(:notthatone, :hello)
+    end
+
+    def test_raw_data
+      markable = MyMarkableObject.new
+      assert_equal({}, markable.raw_data)
+      markable[:hello] = "world"
+      markable[:foo] = "bar"
+      assert_equal({:hello => "world", :foo => "bar"}, markable.raw_data)
+    end
   end
 end