+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.6.0 / FIX ME
 
+* Added Markable#marks and Markable#raw_data
 * Generalized the decoration algorithm to work backwards as well as forwards as well as
   accepting an output Hash-like object, defaulting to true state decorations.
 * Generalized the DFA equivalence algorithm under TransitionSystem::Equivalence.